+++ Assets/Shader/Standard Mobile/Standard_VC.shader
@@ -157,7 +157,7 @@ Shader "Vita/Standard Mobile"
 			
 			struct v2f {
 				V2F_SHADOW_CASTER;
-				float2  uv : TEXCOORD0;
+				half2  uv : TEXCOORD0;
 				UNITY_VERTEX_OUTPUT_STEREO
 			};
 			struct appdata {
@@ -180,7 +180,7 @@ Shader "Vita/Standard Mobile"
 				return o;
 			}
             
-			float4 frag( v2f i ) : SV_Target
+			half4 frag( v2f i ) : SV_Target
 			{
 			
 				fixed4 texcol = tex2D( _MetallicGlossMap, i.uv );
